+++ src/components/button/Delete/Delete.cy.tsx
@@ -1,0 +1,49 @@
+import React from 'react';
+import { CssBaseline, ThemeProvider } from '@mui/material';
+import { THEME_DARK, THEME_LIGHT } from '@ska-telescope/ska-gui-components';
+import theme from '../../../services/theme/theme';
+import DeleteButton from './Delete';
+
+const THEME = [THEME_DARK, THEME_LIGHT];
+const TOOLTIP = 'Tooltip';
+
+function mounting(theTheme: any, disabled: boolean) {
+  cy.viewport(2000, 1000);
+  cy.mount(
+    <ThemeProvider theme={theme(theTheme)}>
+      <CssBaseline />
+      <DeleteButton
+        action={cy.stub().as('action')}
+        disabled={disabled}
+        primary
+        testId="testId"
+        title="BUTTON"
+        toolTip={TOOLTIP}
+      />
+    </ThemeProvider>
+  );
+}
+
+function validateClick() {
+  // cy.get('[data-testid="testId"]').click({ multiple: true });
+}
+
+function validateToolTip() {
+  // cy.get('[data-testid="testId"]').trigger('mouseover');
+  // cy.contains(TOOLTIP).should('be.visible');
+}
+
+describe('<DeleteButton />', () => {
+  for (const theTheme of THEME) {
+    it(`Theme ${theTheme}: Renders (enabled)`, () => {
+      mounting(theTheme, true);
+      validateClick();
+      validateToolTip();
+    });
+    it(`Theme ${theTheme}: Renders (disabled)`, () => {
+      mounting(theTheme, false);
+      validateClick();
+      validateToolTip();
+    });
+  }
+});