+++ src/utils/helpers.ts
@@ -20,7 +20,16 @@ export const countWords = (text: string) => {
         .filter(Boolean).length;
 };
 
-export const displayUnits = (inUnits: string) => (inUnits === 'arcsecs2' ? 'arcsecs\xb2' : inUnits);
+export const presentUnits = (inUnits: string) => {
+  switch (inUnits) {
+    case 'arcsec2':
+      return 'arcsec\xb2';
+    case 'arcsecs2':
+      return 'arcsecs\xb2';
+    default:
+      return inUnits;
+  }
+};
 
 export const helpers = {
   validate: {